requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==3.6.4.0
boto3==1.38.36
botocore==1.38.36
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.2
click==8.1.3
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.3
daphne==4.2.0
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dj-database-url==2.1.0
dramatiq==1.18.0
drf-yasg==1.21.10
exceptiongroup==1.3.0
firebase-admin==6.8.0
git-filter-repo==2.47.0
google-api-core==2.25.0
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
googlemaps==4.10.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==21.2.0
h11==0.16.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jmespath==1.0.1
kombu==5.5.3
msgpack==1.1.0
mysqlclient==2.2.7
outcome==1.3.0.post0
packaging==25.0
pillow==11.2.1
prometheus_client==0.22.0
promise==2.3
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
psycopg2-binary>=2.9,<3.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytube==15.0.0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.3
rsa==4.9.1
s3transfer==0.13.0
selenium==4.33.0
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
text-unidecode==1.3
tomli==2.2.1
trio==0.30.0
trio-websocket==0.12.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
whitenoise==6.9.0
wsproto==1.2.0
yt-dlp==2025.5.22
zope.interface==7.2